Resolve some Darwin compile warnings. (dm)
git-svn-id: svn://mielke.cc/main/brltty@2353 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/sys_shlib_dyld.h
@@ -24,7 +24,7 @@ logDyldError (const char *action) {
   const char *file;
   const char *message;
   NSLinkEditError(&errors, &number, &file, &message);
-  LogPrint(LOG_ERR, "%.*s", strlen(message)-1, message);
+  LogPrint(LOG_ERR, "%.*s", (int)(strlen(message)-1), message);
 }
 
 void *
